# Add first & last_id segments to collection cache_version