--- a/activerecord/test/fixtures/movies.yml
+++ b/activerecord/test/fixtures/movies.yml
@@ -5,3 +5,7 @@ first:
 second:
   movieid: 2
   name: Gladiator
+
+second:
+  movieid: 3
+  name: Ghostbusters