--- a/activerecord/test/models/bagel.rb
+++ b/activerecord/test/models/bagel.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Bagel < ActiveRecord::Base
+end