@jonathanhefner 2019-11-15
Thank you for attacking this issue!  :pray:  :tada:  Unfortunately, there is a counter-case for this approach which appears in #34408.  In #34408, the last page of results changes from `[7, 8, 9]` to `[6, 7, 9]`, but the cache key would remain `"3-12:09-9"` when using this last ID approach.  However, if you used both first and last ID, the cache key would change appropriately.  So the question is: are there also counter-cases that aren't solved by first and last ID (and count and timestamp)?

It seems like there should be, but I can't think of one.  Here are some scenarios:

* record deleted
  * that record's page => last ID or count changes
  * previous pages => no change necessary
  * subsequent pages => last ID or count changes
* record inserted
  * that record's page => max timestamp changes
  * previous pages => no change necessary
  * subsequent pages => last ID or count changes
* record updated / stayed on same results page
  * that record's page => max timestamp changes
  * previous pages => no change necessary
  * subsequent pages => no change necessary
* record updated / moved upward in result pages
  * that record's new page => max timestamp changes
  * that record's old page => first ID changes
  * pages in between => first and last ID change
  * other pages => no change necessary
* record updated / moved downward in result pages
  * that record's new page => max timestamp changes
  * that record's old page => last ID changes
  * pages in between => first and last ID change
  * other pages => no change necessary

Are all of these correct?  And are there cases where combining two or more scenarios might break the solution?

Also, since performance was a concern for #21503, have you benchmarked this approach?