@alipman88 2019-11-17
Thanks for identifying my oversight and proposing a more robust solution, jonathanhefner!

I believe first + last ID does the trick. Consider any page of results:

* If any combination of transactions (inserts/updates/deletes) increases or decreases the number of rows before a page's first ID, the first ID will have shifted when recalculating the cache_version.

* If any combination of transactions changes the number of rows between a page's first ID and last ID, when recalculating the cache version either one or the other will have shifted, or the collection size will have shrunk.

* If any combination of transactions modifies records contained in a page without changing its first ID or last ID, the collection's max updated_at timestamp will have increased.

I was remiss in not including benchmarks originally. Currently, the time required to calculate a cache_version generally appears to grow linearly in regard to the size of a collection. This is because in order to calculate the maximum updated_at column from a subquery of N rows, the database needs to scan through all N rows. (The exception is when the subquery itself isn't efficient, e.g. it's sorted by a non-indexed column - in this case calculating the cache_version will be slow regardless of the algorithm used to do so.)

I believe that in terms of computational complexity/big O notation, both the current cache_key system SQL queries and all the proposed solutions (including the MD5 checksum) run in linear time relative to collection size/limit - so none of them are "horrible."

When including both the first and last ID, I found that using SQL common table expressions slightly improved on my original approach of SQL window functions.

For the purposes of benchmarking, I used a Heroku Postgres DB (standard-0 tier, 4 GB RAM), inserted 2 million records, and with collection sizes/limits ranging from one hundred to one million, benchmarked the time it took for each cache_version algorithm/query to run eight times (four different offsets, ascending & descending sort orders).

The full script and results are at the bottom, but in quick summary: on large collections, the first and last ID approach takes about 50% longer than the current approach, and the MD5 checksum approach takes about 290% longer than the current approach:

https://docs.google.com/spreadsheets/d/1VEH6Rq2lN7nyNCwbVNC5lnOs2TKa1R8Vgh9Vc3_UdcY/edit#gid=742319261

I'm still considering how to elegantly capture the first ID in MySQL - once I feel I've figured that out, I'll update my pull request, add updated benchmarks, and let folks know. (I think it might be best to keep this pull request open so the conversation stays in one place, but the Rails team can feel free to close it if they'd prefer a fresh PR.)

Benchmarking script:

```
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  git_source(:github) { |repo| "https://github.com/#{repo}.git" }

  gem "rails", github: "rails/rails"
  gem "benchmark-ips"
  gem "pg"
end

require "active_record"
require "logger"

ActiveRecord::Base.collection_cache_versioning = true

ActiveRecord::Base.establish_connection({
  adapter: "postgresql",
  url: "redacted"
})

ActiveRecord::Schema.define do
  create_table :users, if_not_exists: true do |t|
    t.string :name
    t.timestamps
  end
end

class User < ActiveRecord::Base
end

# Seed database with one million records
if User.first.nil?
  r = 2_000_000
  t = 5.years.ago.to_i
  now = Time.now.to_i
  pct = 0.0

  puts "initializing test records"

  (1..r).map { |n| {
    id: n,
    name: SecureRandom.alphanumeric(5 + rand(12)),
    created_at: Time.at(t += (now - t) / (1 + r - n)),
    updated_at: Time.at(rand(t..now+1))
  } }.each_slice(20_000) { |slice|
    puts "seeding BD" if pct == 0.0
    User.insert_all(slice)
    puts "#{pct+=1}%"
  }
  
  records.each_slice(20_000) { |slice| User.insert_all(slice) ; puts "#{pct+=1}%" }
end

Benchmark.ips do |benchmark|
  benchmark.config({warmup: 0, time: 60})

  limits = (0..1_000_000).step(100_000).to_a
  limits [0] = 100
  orders = ["ASC", "DESC"]
  offsets = (0..1_000_000).step(250_000)

  limits.each do |limit|
    benchmark.report("current cache_version, limit #{limit}") do
      orders.each do |order|
        offsets.each do |offset|
          ActiveRecord::Base.connection.select_one %(
            SELECT
              COUNT(*) AS size,
              MAX(updated_at) AS timestamp
            FROM (SELECT updated_at FROM users ORDER BY id #{order} LIMIT #{limit} OFFSET #{offset}) q
          )
        end
      end
    end

    benchmark.report("common table expressions, limit #{limit}") do
      orders.each do |order|
        offsets.each do |offset|
          ActiveRecord::Base.connection.select_one %(
            WITH
              q1 AS (SELECT id, updated_at FROM users ORDER BY id #{order} LIMIT #{limit} OFFSET #{offset}),
              q2 AS (SELECT COUNT(*) AS size, MAX(updated_at) AS timestamp FROM q1)
            SELECT
              q2.size,
              q2.timestamp,
              (SELECT id FROM q1 LIMIT 1) AS fst_id,
              (SELECT id FROM q1 LIMIT 1 OFFSET q2.size - 1) AS lst_id
            FROM q2
          )
        end
      end
    end

    benchmark.report("cache_version w/ MD5 checksum, limit #{limit}") do
      orders.each do |order|
        offsets.each do |offset|
          ActiveRecord::Base.connection.select_one %(
            SELECT
              COUNT(*) AS size,
              MAX(updated_at) AS timestamp,
              SUM(18446744073709551616 * ('x' || LEFT(MD5(id::text), 16))::bit(64)::bigint::decimal(58) +
                ('x' || RIGHT(MD5(id::text), 16))::bit(64)::bigint::decimal(58)) AS checksum
            FROM (SELECT id, updated_at FROM users ORDER BY id #{order} LIMIT #{limit} OFFSET #{offset}) q
          )
        end
      end
    end
  end
end
```

Results:
```
Calculating -------------------------------------
current cache_version, limit 100
                          0.446  (± 0.0%) i/s -     26.000  in  60.715002s
common table expressions, limit 100
                          0.498  (± 0.0%) i/s -     29.000  in  60.378481s
cache_version w/ MD5 checksum, limit 100
                          0.446  (± 0.0%) i/s -     26.000  in  60.295986s
current cache_version, limit 100000
                          0.402  (± 0.0%) i/s -     24.000  in  61.291920s
common table expressions, limit 100000
                          0.358  (± 0.0%) i/s -     22.000  in  62.967460s
cache_version w/ MD5 checksum, limit 100000
                          0.171  (± 0.0%) i/s -     11.000  in  65.586271s
current cache_version, limit 200000
                          0.312  (± 0.0%) i/s -     19.000  in  62.331424s
common table expressions, limit 200000
                          0.255  (± 0.0%) i/s -     15.000  in  60.186287s
cache_version w/ MD5 checksum, limit 200000
                          0.126  (± 0.0%) i/s -      8.000  in  63.850849s
current cache_version, limit 300000
                          0.339  (± 0.0%) i/s -     20.000  in  61.407844s
common table expressions, limit 300000
                          0.210  (± 0.0%) i/s -     13.000  in  62.433477s
cache_version w/ MD5 checksum, limit 300000
                          0.096  (± 0.0%) i/s -      6.000  in  62.333182s
current cache_version, limit 400000
                          0.298  (± 0.0%) i/s -     17.000  in  60.079039s
common table expressions, limit 400000
                          0.176  (± 0.0%) i/s -     11.000  in  62.762356s
cache_version w/ MD5 checksum, limit 400000
                          0.082  (± 0.0%) i/s -      5.000  in  61.180664s
current cache_version, limit 500000
                          0.237  (± 0.0%) i/s -     14.000  in  60.591729s
common table expressions, limit 500000
                          0.162  (± 0.0%) i/s -     10.000  in  62.121001s
cache_version w/ MD5 checksum, limit 500000
                          0.066  (± 0.0%) i/s -      4.000  in  60.279141s
current cache_version, limit 600000
                          0.228  (± 0.0%) i/s -     14.000  in  62.951441s
common table expressions, limit 600000
                          0.148  (± 0.0%) i/s -      9.000  in  61.489058s
cache_version w/ MD5 checksum, limit 600000
                          0.058  (± 0.0%) i/s -      4.000  in  68.884699s
current cache_version, limit 700000
                          0.191  (± 0.0%) i/s -     12.000  in  63.534652s
common table expressions, limit 700000
                          0.141  (± 0.0%) i/s -      9.000  in  63.893693s
cache_version w/ MD5 checksum, limit 700000
                          0.050  (± 0.0%) i/s -      3.000  in  60.529919s
current cache_version, limit 800000
                          0.178  (± 0.0%) i/s -     11.000  in  63.929976s
common table expressions, limit 800000
                          0.133  (± 0.0%) i/s -      8.000  in  60.379705s
cache_version w/ MD5 checksum, limit 800000
                          0.046  (± 0.0%) i/s -      3.000  in  64.997166s
current cache_version, limit 900000
                          0.191  (± 0.0%) i/s -     12.000  in  64.194715s
common table expressions, limit 900000
                          0.121  (± 0.0%) i/s -      8.000  in  66.573088s
cache_version w/ MD5 checksum, limit 900000
                          0.042  (± 0.0%) i/s -      3.000  in  71.590107s
current cache_version, limit 1000000
                          0.227  (± 0.0%) i/s -     14.000  in  62.149652s
common table expressions, limit 1000000
                          0.119  (± 0.0%) i/s -      8.000  in  67.486037s
cache_version w/ MD5 checksum, limit 1000000
                          0.038  (± 0.0%) i/s -      3.000  in  78.325545s
```